// src/types/CustomStyleTheme.d.ts
type CustomStyleTheme = {
  colors?: {
    primary?: string;
    secondary?: string;
  },
  fonts?: {
    headingOne?: string;
    headingTwo?: string;
    headingThree?: string;
    body?: string;
  },
  breakpoints?: {
    mobile?: string;
    tablet?: string;
    desktop?: string;
  },
}

export default CustomStyleTheme;